requirements.txt
python-dotenv==1.1.0
neo4j-graphrag[openai]==1.7.0